notification-api/requirements.txt
Flask==1.0
requests==2.21.0
pymongo==3.8.0
flask_pymongo==2.2.0
python-dotenv==0.10.1
flask-expects-json==1.3.1
pika==1.0.1
Flask-Testing==0.6.2
